tests: split test plugins with response and _filter
remove all tests that assume a single plugin can have both.
