+++ spec/02-integration/05-proxy/24-buffered_spec.lua
@@ -143,29 +143,29 @@ for _, strategy in helpers.each_strategy() do
         assert.equal("yes", res.headers["Modified"])
       end)
 
-      it("header can be set from upstream response body on response phase", function()
-        local res = proxy_client:get("/3/status/235")
-        local body = assert.res_status(235, res)
-        assert.equal(md5(body), res.headers["MD5"])
+--       it("header can be set from upstream response body on response phase", function()
+--         local res = proxy_client:get("/3/status/235")
+--         local body = assert.res_status(235, res)
+--         assert.equal(md5(body), res.headers["MD5"])
+--
+--         local res = proxy_ssl_client:get("/3/status/236")
+--         local body = assert.res_status(236, res)
+--         assert.equal(md5(body), res.headers["MD5"])
+--       end)
 
-        local res = proxy_ssl_client:get("/3/status/236")
-        local body = assert.res_status(236, res)
-        assert.equal(md5(body), res.headers["MD5"])
-      end)
-
-      it("header can be set from upstream response body and body can be modified on response phase", function()
-        local res = proxy_client:get("/4/status/237")
-        local body = assert.res_status(237, res)
-        local json = cjson.decode(body)
-        assert.equal(true, json.modified)
-        assert.equal("yes", res.headers["Modified"])
-
-        local res = proxy_ssl_client:get("/4/status/238")
-        local body = assert.res_status(238, res)
-        local json = cjson.decode(body)
-        assert.equal(true, json.modified)
-        assert.equal("yes", res.headers["Modified"])
-      end)
+--       it("header can be set from upstream response body and body can be modified on response phase", function()
+--         local res = proxy_client:get("/4/status/237")
+--         local body = assert.res_status(237, res)
+--         local json = cjson.decode(body)
+--         assert.equal(true, json.modified)
+--         assert.equal("yes", res.headers["Modified"])
+--
+--         local res = proxy_ssl_client:get("/4/status/238")
+--         local body = assert.res_status(238, res)
+--         local json = cjson.decode(body)
+--         assert.equal(true, json.modified)
+--         assert.equal("yes", res.headers["Modified"])
+--       end)
     end)
   end)
 end
