+++ spec/fixtures/custom_plugins/kong/plugins/enable-buffering-response/schema.lua
@@ -1,0 +1,23 @@
+return {
+  name = "enable-buffering",
+  fields = {
+    {
+      config = {
+        type = "record",
+        fields = {
+          {
+            phase = {
+              type = "string",
+              default = "header_filter",
+            },
+          },
+          {
+            mode = {
+              type = "string",
+            },
+          },
+        },
+      },
+    },
+  },
+}
