+++ spec/fixtures/custom_plugins/kong/plugins/enable-buffering/handler.lua
@@ -12,22 +12,22 @@ function EnableBuffering:access()
 end
 
 
-function EnableBuffering:response(conf)
-  if conf.phase == "response" then
-    if conf.mode == "modify-json" then
-      local body = assert(kong.service.response.get_body())
-      body.modified = true
-      return kong.response.exit(kong.service.response.get_status(), body, {
-        Modified = "yes",
-      })
-    end
-
-    if conf.mode == "md5-header" then
-      local body = kong.service.response.get_raw_body()
-      kong.response.set_header("MD5", ngx.md5(body))
-    end
-  end
-end
+-- function EnableBuffering:response(conf)
+--   if conf.phase == "response" then
+--     if conf.mode == "modify-json" then
+--       local body = assert(kong.service.response.get_body())
+--       body.modified = true
+--       return kong.response.exit(kong.service.response.get_status(), body, {
+--         Modified = "yes",
+--       })
+--     end
+--
+--     if conf.mode == "md5-header" then
+--       local body = kong.service.response.get_raw_body()
+--       kong.response.set_header("MD5", ngx.md5(body))
+--     end
+--   end
+-- end
 
 
 function EnableBuffering:header_filter(conf)
